communication-channel/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.19
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.33
MarkupSafe==0.23
Pillow==3.1.0
protobuf==2.6.1
pycrypto==2.6.1
python-axolotl==0.1.7
python-axolotl-curve25519==0.1
python-dateutil==2.4.2
pytz==2015.7
six==1.10.0
Werkzeug==0.11.3
wheel==0.24.0
yowsup2==2.4.48
